% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylToSNP.R
\name{plotPotentialSNPs}
\alias{plotPotentialSNPs}
\title{Plot distribution of beta values for array probes that were identified as SNPs}
\usage{
plotPotentialSNPs(x, betas, horizontal = TRUE)
}
\arguments{
\item{x}{a data frame with probes identified as SNPs by MethylToSNP()}

\item{betas}{A matrix or a data frame or an GenomicRatioSet, GenomicMethylSet, MethylSet, or RatioSet object (see minfi package)}

\item{horizontal}{plot orientation (horizontal=TRUE by default)}
}
\description{
Plot distribution of beta values for array probes that were identified as SNPs
}
\examples{
x <- MethylToSNP(betas)
x <- plotPotentialSNPs(x, data)
}
