% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylToSNP.R
\name{plotProbes}
\alias{plotProbes}
\title{Plot distribution of beta values for array probes}
\usage{
plotProbes(betas, horizontal = TRUE)
}
\arguments{
\item{betas}{A matrix or a data frame or an GenomicRatioSet, GenomicMethylSet, MethylSet, or RatioSet object (see minfi package)}

\item{horizontal}{plot orientation (horizontal=TRUE by default)}
}
\description{
Plot distribution of beta values for array probes
}
\examples{
x <- MethylToSNP(betas)
x <- plotPotentialSNPs(x, data)
}
